pyproject.toml
[project]
name = "live-translator"
version = "0.1.0"
description = "Live bidirectional RU↔EN speech-to-speech translator for macOS"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "sounddevice>=0.4.6",
    "faster-whisper>=1.1.0",
    "ctranslate2>=4.0",
    "sentencepiece>=0.2.0",
    "piper-tts>=1.2.0",
    "requests>=2.31",
    "webrtcvad-wheels>=2.0.11",
    "rumps>=0.4.0",
    "pynput>=1.7.6",
    # Web server dependencies
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "aiortc>=1.8.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
]

[project.scripts]
translator = "src.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
